requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
idna==3.4
Pillow==9.2.0
psycopg2==2.9.4
python-decouple==3.6
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.2
typing_extensions==4.4.0
tzdata==2022.2
urllib3==1.26.14
webencodings==0.5.1
